web/requirements.txt
alabaster==0.7.12
alembic==1.0.3
aniso8601==4.0.1
apns3==1.0.0
asn1crypto==0.24.0
astroid==1.4.9
attrs==18.2.0
Babel==2.6.0
bcrypt==3.1.4
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.2
cryptography==2.4.1
decorator==4.3.0
docutils==0.14
ecdsa==0.13
fabric==2.4.0
flake8==2.5.1
Flask==1.0.2
Flask-Auth==0.85
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.4
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.4.2
GeoAlchemy2==0.5.0
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
haversine==1.0.2
httplib2==0.12.0
idna==2.7
imagesize==1.1.0
invoke==1.2.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
lint==1.2.1
Mako==1.0.7
manage.py==0.2.10
MarkupSafe==1.1.0
mccabe==0.3.1
paramiko==2.4.2
pbr==5.1.1
pep8==1.7.1
Pillow==5.3.0
pluggy==0.6.0
py==1.7.0
py-postgresql==1.2.1
pyasn1==0.4.4
pycparser==2.19
pycrypto==2.6.1
pyflakes==1.0.0
Pygments==2.2.0
pylint==1.6.1
PyNaCl==1.3.0
PySocks==1.6.8
pytest==3.4.0
pytest-cov==2.5.1
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
requests==2.20.1
setuptools-lint==0.5.2
six==1.11.0
smmap2==2.0.5
snowballstemmer==1.2.1
Sphinx==1.6.6
sphinxcontrib-httpdomain==1.6.0
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.14
sqlparse==0.2.4
stripe==2.12.0
Tempita==0.5.2
tox==2.1.1
twilio==5.6.0
urllib3==1.24.1
virtualenv==15.0.2
Werkzeug==0.14.1
wrapt==1.10.11
